in vivo: initialize tumor cells and Tcell subsets
in vitro started with only a few cells. Here we start with 25-95% of pixels with tumor cells
Blast cells in the bone marrow are typically between 25 and 95% in patients with ALL.
//      Source: Emadi, A., & York Law, J. (2023). Acute lymphoblastic leukemia (ALL). In Merck Manual (Professional Edition). Merck & Co. https://www.merckmanuals.com/professional/hematology-and-oncology/leukemias/acute-lymphoblastic-leukemia-all
//  31%
//      Mahmood, Khalid, Muhammad Ubaid, and Syeda Taliya Rizvi. "Multiple Osteolytic Lesions Causing Hypercalcemia: A Rare Presentation of Acute Lymphoblastic Leukemia." Case Reports in Medicine, vol. 2017, 2017, Article 2347810. Wiley Online Library, https://doi.org/10.1155/2017/2347810.
and
2.5% CD8 (https://pubmed.ncbi.nlm.nih.gov/15723813/)
with 80% bystander TIL/inactive TIL (https://pubmed.ncbi.nlm.nih.gov/38744277/)
and 20% active T cells. This is in line with paper in ALL BM finding that 20% of CD8 T cells expressed PD-1 (https://www.nature.com/articles/s41375-018-0360-1)